test: config update
