+++ package.json
@@ -67,17 +67,15 @@
     "lint": "npm run lint:tsc && eslint --ext .vue,.js,.ts,.tsx ./src --max-warnings 0",
     "lint:tsc": "tsc --emitDeclarationOnly",
     "test": "npm run test:unit && npm run test:snap",
-    "test:demo": "node src/_common/test/script/generate-demo-test.js",
-    "test:unit": "vitest run --config site/vite.config.js",
-    "test:update": "vitest run --config site/vite.config.js -u && npm run test:snap-update",
-    "test:unit-dev": "vitest --config site/vite.config.js",
-    "test:unit-gui": "vitest --config site/vite.config.js --ui",
-    "test:unit-coverage": "vitest run --config site/vite.config.js --coverage",
-    "test:snap": "cross-env NODE_ENV=test-snap vitest run --config site/vite.config.js",
-    "test:snap-update": "cross-env NODE_ENV=test-snap vitest run --config site/vite.config.js -u",
+    "test:unit": "vitest run --config test/unit/vitest.config.js",
+    "test:update": "vitest run --config test/unit/vitest.config.js -u && npm run test:snap-update",
+    "test:unit-dev": "vitest --config test/unit/vitest.config.js",
+    "test:unit-gui": "vitest --config test/unit/vitest.config.js --ui",
+    "test:unit-coverage": "vitest run --config test/unit/vitest.config.js --coverage",
+    "test:snap": "cross-env NODE_ENV=test-snap vitest run --config test/unit/vitest.config.js",
+    "test:snap-update": "cross-env NODE_ENV=test-snap vitest run --config test/unit/vitest.config.js -u",
     "test:e2e": "cypress run --config-file script/cypress/cypress.config.js --component",
     "test:e2e-gui": "cypress open --config-file script/cypress/cypress.config.js",
-    "changelog": "node script/generate-changelog.js",
     "robot": "publish-cli robot-msg",
     "prepare": "husky install"
   },
