+++ test/cypress/support/component.js
