--- test/e2e/menu/menu.spec.js
@@ -1,13 +1,0 @@
-describe('测试Menu组件', () => {
-  beforeEach(() => {
-    cy.visit('/#/components/menu');
-  });
-  it('测试head-menu', () => {
-    cy.get('.tdesign-menu-head-menu')
-      .find('div')
-      .first()
-      .should((ele) => {
-        expect(ele).to.have.class('t-head-menu');
-      });
-  });
-});
